pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "delivery*", "sources*", "utils*"]

[project]
name = "agentmesh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "deepagents>=0.4.11",
    "duckduckgo-search>=8.1.1",
    "feedparser>=6.0.12",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.3.5",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langgraph-checkpoint>=4.0.1",
    "langgraph-sdk>=0.3.11",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "tavily-python>=0.7.23",
]
